merge master

+++ packages/@romejs/core/test-worker/TestWorkerRunner.ts
@@ -28,7 +28,7 @@ import {
   convertTransportFileReference,
 } from '../common/types/files';
 import {GlobalTestOptions} from '@romejs/test';
-import {createAbsoluteFilePath} from '@romejs/path';
+import {createAbsoluteFilePath, AbsoluteFilePath} from '@romejs/path';
 
 const MAX_RUNNING_TESTS = 20;
 
@@ -39,9 +39,11 @@ export default class TestWorkerRunner {
     this.file = convertTransportFileReference(opts.file);
     this.options = opts.options;
     this.bridge = bridge;
+    this.projectFolder = createAbsoluteFilePath(opts.projectFolder);
+
     this.snapshotManager = new SnapshotManager(
+      this,
       createAbsoluteFilePath(opts.file.real),
-      opts.options.updateSnapshots,
     );
 
     this.hasFocusedTest = false;
@@ -59,6 +61,7 @@ export default class TestWorkerRunner {
   hasFocusedTest: boolean;
 
   bridge: TestWorkerBridge;
+  projectFolder: AbsoluteFilePath;
   file: FileReference;
   options: TestRunnerOptions;
   snapshotManager: SnapshotManager;
@@ -81,6 +84,13 @@ export default class TestWorkerRunner {
     return {
       __ROME__TEST_OPTIONS__: testOptions,
     };
+  }
+
+  async emitDiagnostic(diagnostic: PartialDiagnostic) {
+    await this.bridge.testError.call({
+      ref: undefined,
+      diagnostic,
+    });
   }
 
   // execute the test file and discover tests
@@ -161,7 +171,8 @@ export default class TestWorkerRunner {
 
     let diagnostic: PartialDiagnostic = deriveDiagnosticFromError({
       error: opts.error,
-      category: testName,
+      category: 'tests/failure',
+      label: testName,
       filename,
       cleanFrames(frames) {
         // TODO we should actually get the frames before module init and do it that way
@@ -254,12 +265,13 @@ export default class TestWorkerRunner {
       };
     });
 
-    const api = new TestAPI(
+    const api = new TestAPI({
+      file: this.file,
       testName,
       onTimeout,
-      this.snapshotManager,
-      this.options,
-    );
+      snapshotManager: this.snapshotManager,
+      options: this.options,
+    });
 
     try {
       const res = callback(api);
@@ -286,11 +298,23 @@ export default class TestWorkerRunner {
     }
   }
 
-  async runTests(): Promise<void> {
+  async run(): Promise<void> {
     const promises: Set<Promise<void>> = new Set();
 
+    const {foundTests} = this;
+    if (foundTests.size === 0) {
+      this.bridge.testError.send({
+        ref: undefined,
+        diagnostic: {
+          filename: this.file.uid,
+          message: 'No tests declared in this file',
+          category: 'tests/noneDeclared',
+        },
+      });
+    }
+
     // Execute all the tests
-    for (const [testName, {options, callback}] of this.foundTests) {
+    for (const [testName, {options, callback}] of foundTests) {
       if (callback === undefined) {
         continue;
       }
@@ -304,14 +328,19 @@ export default class TestWorkerRunner {
       });
 
       const promise = this.runTest(testName, callback);
-      promise.then(() => {
-        promises.delete(promise);
-      });
-      promises.add(promise);
 
-      // if there's 5 promises, then wait for one of them to finish
-      if (promises.size > MAX_RUNNING_TESTS) {
-        await Promise.race(Array.from(promises));
+      if (this.options.syncTests) {
+        await promise;
+      } else {
+        promise.then(() => {
+          promises.delete(promise);
+        });
+        promises.add(promise);
+
+        // if there's 5 promises, then wait for one of them to finish
+        if (promises.size > MAX_RUNNING_TESTS) {
+          await Promise.race(Array.from(promises));
+        }
       }
     }
 
@@ -323,7 +352,7 @@ export default class TestWorkerRunner {
   }
 
   async emitFoundTests() {
-    const promises: Array<Promise<void>> = [];
+    const tests = [];
 
     for (const [testName, {callback, options}] of this.foundTests) {
       let isSkipped = callback === undefined;
@@ -331,30 +360,21 @@ export default class TestWorkerRunner {
         isSkipped = true;
       }
 
-      promises.push(
-        this.bridge.testFound.call({
-          ref: {
-            filename: this.file.real.join(),
-            testName,
-          },
-          isSkipped,
-        }),
-      );
+      tests.push({
+        ref: {
+          filename: this.file.real.join(),
+          testName,
+        },
+        isSkipped,
+      });
     }
 
-    await Promise.all(promises);
+    await this.bridge.testsFound.call(tests);
   }
 
-  async run(): Promise<void> {
+  async wrap(callback: () => Promise<void>): Promise<void> {
     try {
-      // Setup
-      await this.snapshotManager.load();
-      await this.discoverTests();
-      await this.emitFoundTests();
-
-      // Execute
-      this.lockTests();
-      await this.runTests();
+      await callback();
     } catch (err) {
       const diagnostics = getDiagnosticsFromError(err);
       if (diagnostics === undefined) {
@@ -379,5 +399,17 @@ export default class TestWorkerRunner {
         }
       }
     }
+  }
+
+  async prepare(): Promise<void> {
+    return this.wrap(async () => {
+      // Setup
+      await this.snapshotManager.load();
+      await this.discoverTests();
+      await this.emitFoundTests();
+
+      // Execute
+      this.lockTests();
+    });
   }
 }
